backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
channels
channels-redis
daphne
Pillow
redis
python-decouple
psycopg2-binary
gunicorn
whitenoise
celery
razorpay
cloudinary
pusher
python-dotenv